Added logic for wallet related stuff

+++ src/main/java/com/paypal/walletms/pojo/WalletHoldRequest.java
@@ -1,9 +1,5 @@
 package com.paypal.walletms.pojo;
 
-import java.util.List;
-
-import com.paypal.walletms.error.ErrorDetail;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -15,6 +11,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class WalletHoldRequest {
 	private Long userId;
-	private Long amount;
+	private Double amount;
 	private String currency;
 }
